feat(parseOperator & parseQuantifier): '{' is parsed as Operator and parseQuantifier correctly advances
This updates parseOperator and parseQuantifier to correctly handle {n} and {n,n} quantifiers.
parseQuantifier needs to advance to next token for the two edge cases where it finds its closing '}' token.
Otherwise, the parser will re-create a character token for that '}' that is really just a part of the quantifier.