fix(compiler): fix regex parsing of back reference characters (#1537) (#1661)
**Named Backreferences**
The regex parser always emitted an unrecoverable `UNCLOSED_NAMED_CAPTURE` diagnostic for named back
references because the `JSRegExpNumericBackReference` returned by `tokenizeEscaped` removed the `<>` from the parsed name, but the `parseCharacter` function stripped the `<` and asserted that the name ends with a `>`.

This diff changes `tokenizeEscaped` to return the name with the `<>`. The `parseCharacter` function then asserts for the closing `>` and strips `<>` from the name.

**Numeric Backreferences**
The regex parser incorrectly parsed numeric backreferences in regular expressions as unicode characters. e.g. the backreference `\1` in the `/f(o)\1/;` expression gets parsed as `JSRegExpCharacter {value: '\x01'}` instead of a `JSRegExpNumericBackReference{value: 1, ...}`

This commit adds the parsing of numeric backreferences according to the [ecma script specification](https://tc39.es/ecma262/multipage/text-processing.html#prod-DecimalEscape), treating numbers starting with `1-9` as numeric back-references.

**Other enhancements**

* Emits a diagnostic if a regex contains multiple named captures with the same name
* Emits a diagnostic if a regex contains a backreference to a non-existing named group