feat(formatter): Add helpers for creating `FormatToken`s (#1690)
This PR introduces various helpers to create `FormatToken`s more conveniently and giving us the flexibility to lower repetitive patterns to the format IR understood by the printer.
Lowering allows us to model an IR that is optimal for the printer and without having to create a custom IR element for every use case.

This PR removes some implicit `Into<FormatToken>` conversions because I believe that the more explicit use of e.g. `token("test")` improves readability and discoverability.

This PR renames some tokens:

* Removes the `Token` postfix for all tokens
* `String` -> `Token` to align with the Rowan naming
* `IfBreak` -> `ConditionalGroupContent` because it only has a meaning inside of groups

This PR intentionally refers to `FormatToken`s as `elements` because we plan to rename `FormatToken` to `FormatElement` as part of #1688.