feature(formatter): format ASTs with missing tokens or nodes (#1707)
Adds support for formatting ASTs with missing tokens or nodes which is the first step on our journey to support (partially) formatting source files with arbitrary syntax errors.

The main change is to return an `Option<FormatElement>` instead of a `FormatElement` from `ToFormatElement.to_format_element` if a node or token contains invalid content (e.g. a mandatory token or child node is missing). The parent has then the option to implement some recovery logic, for example inserting the node with its current formatting, or can propagate the formatting error (which most nodes will do). 

This PR does not address the issue where tokens and/or nodes are present in places they aren't expected in a valid syntax tree.