Change the visibility of green nodes to internal
The `Green*` types are an internal optimisation of our CST implementation to achieve node re-use. This detail shouldn't escape the rowan package.

The `cursor` types are another internal abstraction.

Both of these internal structure shouldn't be consumed outside of the rowan crate (Roslyn defines the cursor and green types `internal` as well).

This PR changes the visibility of the `Green*` and `cursor::*` types to `pub::crate`. I had to remove the `replace_with` methods to achieve this. We should revisit the exact signature and behavior of that function when working on the mutation API.

How about multi threading. I believe it be sufficient to just clone the root of the `SyntaxTree` and move it to the new thread rather than passing `GreenNode`s around.