Pre and PostUpdateExpression
Splits out the `PreUpdateExpression` and `PostUpdateExpression` from the `UnaryExpr`.
Separating the two is necessary for implementing fixed slots for children (#1724) because it's otherwise unclear if the operand or operator comes first.

It would have been sufficient to only introduce the `PostUpdateExpression` and model the `PreUpdateExpression` as an `UnaryExpr`. However, I do think that it's more important to keep the symmetry of the Update expressions and the previously existing `is_update()` helpers on the `UnaryExpr` suggest that it's important for passes to know if they're operating on an `Update` expression or not (which is exactly what splitting the nodes gives you).