Unify `JsModule` and `JsScript` into `JsRoot` and parse directives
This PR unifies the `JsModule` and `JsScript` into the `JsRoot` and adds the source type to the `Parse` result (if it's a module or a script).

The main reason for doing so is that we decided to use the exact same grammar for both scripts and modules (we're not preventing import statements inside of scripts). Mainly, to not have to parse module syntax as unknown statements inside of scripts (flagging them as errors should be sufficient).

Unifying modules and scripts further simplifies the API usage because there's now a unified root type.

This PR further refines the parsing of directives and adds directive to our grammar by introducing a new `FunctionBody` node for blocks that are direct children of any function declaration (and, therefore, allow directives).