Remove `Decl` from AST Facade
This PR removes the `Decl` union type because it might be misleading. It could be understood to contain any possible declaration whereas it only contains declarations that are allowed in `export` statements. However, it doesn't include other identifier declarations like: named `function expressions`, `class expressions`, parameters, etc....

We may want to re-introduce a form of `Decl` in the future if we see usage for it.