refactor(rslint_parser): Refine Function Expression/Declaration and arrow function expression AST tree structures (#1786)
Updates the facade for function declarations and function expression to match the grammar defined in #1719. 

* Function declaration
  * Rename from `FnDecl` to `JsFunctionDeclaration`
  * Rename `ParameterList` to `JsParameterList`
  * Introduce new `JsIdentifierBinding` type for `id`
  * Parses `...rest` as `JS_REST_PARAMETER` instead of `JS_REST_PATTERN`.
  * Split out `TsReturnType` for the `:` and the return type
* Function expression
  * Rename from `FnExpr` to `JsFunctionExpression`
  * Otherwise the same as for declaration
* Arrow function expression
  * Rename from `ArrowExpr` to `JsArrowFunctionExpression`
  * Rename `ExprOrBlock` to `JsAnyArrowFunctionBody`
  * Otherwise the same as for declaration.

This PR cleans up some of the function parsing (extracts helpers rather than using boolean params). This improves error reporting for the `func_decl` use case if the name is missing. 

This PR also removes the support for decorators. The parser had some support for decorators but they haven't been part of the grammar. We should re-add them when we decide to properly support them (and add tests)