refactor(rslint_parser): Restructure Object AST Tree strcuture (#1787)
This restructures the AST structure of object expressions and related nodes (members, etc.) and is part of #1725.

* Renames `ObjectExpr` to `JsObjectExpression`
* Renames `props` to `members`
* Renames the `key` of all members to `name`
* Renames `IdentProp` to `JsShorthandPropertyObjectMember` that directly stores a `ident` token
* Renames `LiteralProp` to `JsPropertyMember` and changes the `key` from `Name` to `JsStaticObjectMemberName`
* Introduces `JsGetterObjectMember` (unclear if it should be shared with classes in the future)
* Introduces `JsSetterObjectMember` (unclear if it should be shared with classes in the future)
* Introduces `JsMethodObjectMember` (unclear if it should be shared with classes in the future)
* Renames `SpreadProp` to `JsSpread`

The PR fixes some issues when it comes to `setter/getter` parsing.

The main concern around sharing `Js*Member`s with classes is that objects are more restrictive when it comes to naming because classes support private properties which object members don't. I think the better approach is to introduce a `JsAnyGetter = JsGetterObjectMember | JsGetterClassMember` union that implements helpers to retrieve the shared properties.