feat(rslint_parser): Require explicit complete/abandon calls for markers
This PR adds a drop bomb to the Parser's `Marker` implementation to ensure we always call either `abandon` or `complete` on every marker. 

I was stunned to see that the parser is able to produce a valid tree even with nodes in the tree that were never completed, and, therefore, don't have a `SyntaxKind` set. 

The reason why this works is that the parser sets the kind to `TOMBSTONE` when starting a new node and these get skipped when building the final tree. The problem is, skipped means the children of the `TOMBSTONE`d node get attached to the parent. 

This "bubbling up" of children is problematic, because let's say we have a node that calls into `expr`. The expectation would be that `expr` either produces a `UnknownExpression` or doesn't add any nodes to the parse tree. However, this isn't the case. `expr` may start parsing and then decide that it doesn't know how to continue and it then bails out with the result that all parsed tokens and nodes get attached to the parent. 

Bailing out is fine, but we should be very considerate about what we want to do in those cases, either completing the node with an `UNKNOWN` kind or propagating the error to the parent (by, for example, returning a `Result` type).