Print Missing slots in debug output
We plan to use fixed offsets to access any child of a node. That requires, that the parser inserts "empty slots" if a child is missing (either because it's optional or because of a syntax error).

Our current means of testing the parser are the snapshot tests that use the alternative debug format to assert the tree structure. However, this format currently doesn't contain any output for empty slots.

This PR adds the empty slots to the alternate debug output of SyntaxNodes so that we can verify that the parser inserts missing slots where needed.

This PR further renames `expect` to `required_token` to be more explicit that this will add a missing marker and introduces a `optional_token` for cases where one wants to eat an optional token if present.