Allow missing tokens/nodes in AstSeparatedList
I initially proposed that the `AstSeparatedList` should panic whenever a separator or node is missing. However, that means that the only way for the parser to recover from a missing separator (e.g. `a b`) or from a missing node (`,a`) is to create an empty token (COMMA with "text": "") or an empty element.

Having these empty tokens and nodes breaks with the convention used in the rest of the AST facade where missing required elements return a `SyntaxError::MissingRequiredChild`.

This PR changes the `AstSeparatedListElement` to store the node and separator tokens as `SyntaxResult<T>`s and removes the `panics` in the implementation.

This allows us now to also add the `Ast` snapshot to the failing dir tests.