feat(rslint_parser) Assert AST Test in parser integration test (#1827)
We test our parser by comparing the expected syntax tree with the actual tree produced by the parser. This works well but doesn't give us any guarantee that what the parser produces matches the AST structure. 

This PR removes the derived `Debug` implementation for `AstNode` and replaces it with a generated implementation that prints the children of a node and adds the AST output to the parser's `ok` integration tests.