#1829 introduced a bug where the parser now is recursing forever for some libraries like `Vue`.
The problem seems to be from the change from `SingleTokenParseRecovery` in `pat` to use `ParseRecovery` instead. However, this change is necessary because parameter list may otherwise contain two elements that are not separated by a comma.

This PR adds a somewhat hacky fix since I'm rewriting `patterns` anyway. It reverts `pat` to use `SingleTokenParseRecovery` but takes a snapshot in `parse_formal_param_pat` and rewinds if `pattern` returned `None`.

This fixes most recursing forever issues. There are still some more but they aren't specific to #1829