feat(rslint_parser: Parser Progressing Assertion (#1834)
RSLint used an `oferflow_check` inside of the parser that verified if the parser did less than 50 000 000 steps to parse the current file and otherwise panics, saying that the parser is recursing forever. 

The nicety about this approach is that it doesn't require explicit checks during parsing. However, it has a few downsides:

* It panics for very large file 
* Overflow checks are only necessary in while loops
* The panic happens anytime, but not necessarily in the loop that stopped progressing and the parser goes on and on for a long time until it reached the step limit

This PR introduces a `ParseProgress` struct that allows asserting whatever the parser is making progress. This new struct must be used in every `while` loop.