`or_missing*` helpers for `ParsedSyntax<ConditionalSyntax>`
`ParsedSyntax<ConditionalSyntax>` currently offered no helpers to mark the result as missing and adding an error. This resulted in quiet some cumbersome code.

This PR adds the `or_missing` and `or_missing_with_error` helpers and uses the new helpers on the call sites.

This PR also renames `InvalidParsedSyntax` to `InvalidSyntax` to align it with `ConditionalSyntax`

The tests output changes because we now add the "missing" markers for parameter types and return types.