chor(rslint_parser): Remove list workarounds for trivia/error nodes
The node list had to gracefully handle the case where a list could contain trivia nodes or errors.

This is no longer needed because trivia are now stored inside of tokens and (most) error nodes have been replaced with `Unknown*` nodes that can be casted to `N`. That's why this PR removes the workarounds