fix(rslint_parser): Member of computed assignment targets
A computed assignment target contains two expressions, one for the object and one for the member and the facade didn't manually implement the `member` field. The result (until we use fixed slots) is that both `member` and `object` return the same `JsAnyExpression`.

This went unnoticed because the AST code gen didn't add manual fields to the `Debug` implementation of a node. This diff changes the code gen to also include manual fields to surface such issues in the future.

Adding the `manual` field support to the code gen surfaced some typing mismatches between our ungrammar file and the implementations that I fixed as well.