feature(rslint_parser): Debug print for Unknown* node children (#1866)
The children of `Unknown` nodes are simply a sequence of `SyntaxElement`s. That's why the items currently print without field names nor validation if the CST matches the format of our AST facade.

This PR adds an `AnyNode` union that is a union over all nodes defined in the grammar, allowing to cast any `SyntaxNode` to a Node and, from there, call the correct `Debug` implementation for a node. So that `Unknown` nodes now print their children using their AST representation.