feature(rslint_parser): Class Modifiers
I noticed that the class parsing doesn't conform to our AST structure when working on the automatic AST shape validation (marks nodes as `UNKNOWN` if they don't match the expected AST shape).

The reason for that is that some members have `missing` slots for modifiers they don't support or a `missing` slot for an optional modifier is missing.

I further noticed that the parsing didn't correctly support members like `private()` because we only had special handling for `declare` and `static`.

This PR rewrites the modifiers handling and adds support for all valid member names (including `declare`, `static`, `private`, ...).