refactor(rslint_parser): Specific List nodes (#1875)
Implementing the shape validation described in #1867 requires that it's possible to tell the required shape of a node by just looking at its kind. This hasn't been possible so far for list nodes that all use the same ` LIST` kind. The kind doesn't expose any information of what the expected type of the child nodes is nor if it is a separated or a node list. 

This PR replaces the generic `LIST` kind with specific list types, for example, it introduces the `JsArrayElementList` for the list storing the elements of an array expression. 

A consequence of having explicit `LIST` nodes is that these should now also implement the `AstNode` interface, they are nodes after all. The consequences of doing so are:

* `AstSeparatedList` and `AstNodeList` must be traits that each specific list implement together with the `AstNode` interface
* It's no longer possible to use a `missing` placeholder to represent an empty list because a concrete node is required to implement `AstNode.syntax()`. 

Overall, the change seems to nicely align with how the rest of the AST facade works and makes list nodes less special, they're just regular nodes that on top implement additional methods that allow iterating over their elements. 


The main downside of this change is that it's now required to implement the `AstSeparatedList` and/or `AstNodeList` traits if someone wants to use any of the list methods. I guess, that's just how Rust works and shouldn't be too much of a surprise.

Part of #1725 #1867