Fix assignment to array or object expression members
The tests added in #1893 revealed that our parser incorrectly parses the inner assignment target (`[1, 2][0]`) in `[[1,2][0] = 42] = [23]` as an array assignment target instead of a computed member expression. The same is true when using an object as the left-hand side of a member expression inside an assignment pattern.

This diff fixes this by first parsing the expr and only converting it to an assignment if it is an `array` or `object` expression