feature(rslint_parser): Any Class and Function
It's common that analyzers or AST visitors may want to implement some logic that applies to all classes or functions.

This PR introduces two auxiliary unions `JsAnyClass` and `JsAnyFunction` that expose the common children and demonstrate their use by refactoring the class and function formatting to now use the new unions.

The main motivation for introducing these auxiliary unions now is to demonstrate that adding a new "class" and "function" node for export statements is a viable solution to correctly model that the `id` for these declarations is optional.

This PR further renames `JsClassDeclaration` to `JsClassStatement` and `JsFunctionDeclaration` to `JsFunctionStatement` as agreed on in #1868