fix(rslint_parser): libs parse diagnostics
This PR fixes the remaining diagnostic when parsing the `coverage-libs` libs.

### Labels

The parser tracked all defined labels but never removed them when leaving the labeled statement. That's why the parser reported diagnostics for eligibly duplicated label declarations if any label was re-used anywhere in the program

```js
a: {
  a: {} // not ok
}
a: {} // ok
```

### Variable declarations

The parser generates diagnostics if a `let` or `const` declaration defines the same identifier twice. However, re-declaring the same identifier inside of the initializer is totally fine:

```js
let a = () => { let a = 10; } // this is ok
let a, a; // this isn't
```

### Parse libs as scripts

Parse the libs with `script` to avoid the warning that the `use strict` directive is redundant.