feat(rslint_parser): Avoid cloning parser state (#1925)
The parser has a neat `with_state` function that takes a snapshot of the parser state and restores the state to its previous state when it goes out of scope. However, the downside is that it copies the state at least twice, once inside the `with_state` function and once in the call-site, when constructing the new state which results in a serious performance penalty. 

This PR introduces a new `ChangeParserState` trait that allows making specific changes to the `ParserState` that can later be undone, only keeping a copy of the data they changed (and need to be restored later). Ideally, these changes would have semantic names. However, this PR re-uses the name of the fields.