fix(rslint_parser): Disallow continue for non iteration labelled statements
`continue` statements are only valid inside a loop or if they point to a labelled iteration statement (do, while, for).

This PR changes the `label_set` to track what the kind of the labelled statement is (uses lookahead to determine the kind) and validates that any label used with `continue` points to an iteration.

Furthermore, fixes an issue where we didn't respect the negative lookahead of `let [` for `parse_expression`.