feat(rslint_parser): Parser State Checkpoint
This isn't an issue for most of our state variables because most changes are scoped, automatically restored when exiting the parsing of a section.

However, the `hoisted` and `scoped` variables changes aren't scoped. A variable gets added when its first seen and the change is only undone when entering a new function or block (or class). Thus, re-parsing the same binding after a restore triggers a re-declaration error.

This PR is a preparation for #1929. It adds a `ParserStateCheckpoint` that ensures the `Parser` and its state are correctly rewinded when using checkpoints. The current implementation only adds debug assertions because the current state changes are all scoped, but it provides the infrastructure to rewind some state as well. The additional assertion will hopefully help to catch issues as noticed in #1929 sooner.

Creating a checkpoint must be cheap, that's why this PR changes the `ParserState` to use `IndexMap` that already is a dependency of the LSP crate. An `IndexMap` retains insertion order, thus, restoring to a previous state is as simple as calling `truncate(old_index)` on the map.