fix(rslint_parser): Assignment shorthand prop with initializer (#1953)
There are a few places in our code where it isn't possible to tell ahead of time if a syntax is an assignment or an expression.
The parser tries to first parse the syntax as an expression because expressions are a superset of assignment with the exception that an assignment pattern allows a default initializer for shorthand properties.

This PR gracefully handles the case where an object shorthand property has a default initializer so that it won't drip the parser when parsing an expression that later turns out to be an assignment. It has the nice added benefit that it adds a nice error message for users.