fix(rslint_lexer): Fix lexing of regex literals
Fix remaining 262 tests that incorrectly fail parsing a valid program containing a regex literal.
The lexer uses the `in_expr` state to determine whatever a regexp is valid. This requires, that the lexer correctly tracks in which places an expression is valid. This hasn't been the case for

```
{

}
function // <- smt
```

and some keywords that I added to `before_expr` list.