fix(rslint_parser): Left-recursive sequence expression (#1977)
Fixes the sequence expression parsing to be left recursive. I took this as a chance to rewrite `parse_paren_or_arrow_expr`. It now first tries to parse the arguments as sequence expression but bails out early if it encounters any parameter specific syntax.

The performance should be about the same because the old implementation also rewound to re-parse the arrow function parameters. It just did so spread across multiple places.