fix(rslint_lexer): Invalid unicode escape sequence in tagged template literal (#1985)
Invalid unicode escape sequences are valid inside of tagged template literals, but not template literals. 

This PR suppresses the escape sequence diagnostics for tagged template literals. 

It further fixes an issue with escaped sequences where `validate_escape_sequence` progressed the current position and so did `read_str_literal`. 

It also fixes the span of a diagnostic to be a valid range (and not from pos to 1).