feat(rslint_parser): TS Type Arguments (#2032)
Re-implements parsing of type arguments in call and new expressions

```ts
call<A, B, C>()
```

This PR includes a refactoring that splits the `subscripts` parsing into different methods for member vs call subscripts. I found it rather difficult to reason about the code when all `subscripts` were in a single method. 

Part of #2027