feature(rslint_parser): TS Optional & Definite Members and Variables (#2057)
Implements parsing for optional and definite members and variables

```
let a!: string;
class A {
  a!: string;
  b?: string = "test";
  c?() {}
}
```

This PR introduces the new AST nodes

```
TsAnyVariableAnnotation =
	TsTypeAnnotation
	| TsDefiniteVariableAnnotation

TsDefiniteVariableAnnotation =
	'!'
	type_annotation: TsTypeAnnotation
```

and

```
TsAnyPropertyAnnotation =
	TsTypeAnnotation
	| TsOptionalPropertyAnnotation
	| TsDefinitePropertyAnnotation

TsOptionalPropertyAnnotation =
	'?'
	type_annotation: TsTypeAnnotation?

TsDefinitePropertyAnnotation =
	'!'
	type_annotation: TsTypeAnnotation
```

These model the constraints that type annotations are mandatory if the `!` operator is used. However, the nodes do not model that an initializer clause isn't allowed in combination with `!`.

One reason for this is simply naming. I haven't been able to come up with a good name for such a node.

The second reason is that I believe that an initializer on a definite property shouldn't prevent a bundler/test runner from running the JavaScript file. Thus, the parser must be able to return a valid node if both the `!` and the initializer are present.