chor: GitHub ignore generated files
Changes to the AST create changes in many generated files. These changes can be distracting from the real changes made by the author.

This PR sets the `linguist-generated=true` attribute for the files generated by the codegen so that they are hidden in diffs by default and no longer count to the repo's statistics.

[Source](https://docs.github.com/en/repositories/working-with-files/managing-files/customizing-how-changed-files-appear-on-github)