feat(rslint_parser): Parse ambient `enum` and `type` declarations (#2076)
Implements parsing of enums and type aliases in ambient contexts: 

```ts
declare enum A { X, Y }
declare const enum B { X, Y }
declare type C = A;
```

This PR doesn't implement parsing of enums inside other ambient contexts like:

```ts
declare namespace a {
  enum A { X, Y }
}
```

This PR reuses the `TsEnumStatement` and `TsTypeAliasStatement` and nests them inside of a `TsDeclareStatement`. This has the benefit that we don't need new nodes but comes at the downside that it may feel weird that a `statement` can appear in such a position. 

Another downside is that an analyzer needs to probe the enum's or alias statement's parent to figure out if it's inside of an ambient context or not.