feat(rslint_parser): TS function overload syntax
TypeScript allows function overloading by having a function statement without a body followed by a function statement with the same name and a body

```ts
function test();
function test(a: string | undefined) {}
```

This PR re-uses the `TsDeclareFunction` node (and renames it to `TsDeclareFunctionStatement`) to represent function overloads.
This comes at the downside that `declare` functions now have an optional `async` child even though they can't be async.

The PR regresses the *no-optional pattern parameter for function parameters* it's no longer known upfront if a function is an implementation or a function overload.
It would be possible to use a checkpoint and rewind the parser in case it turns out to be a function overload definition, but I want to create a snapshot in such a hot path and believe that such a check should be part of a second-pass.