refactor: `JsAnyDeclarationClause` (#2098)
This PR introduces a new `JsAnyDeclarationClause` for declarations in non-statement positions. The union variants are the same as `JsAnyDeclaration` with the exception of `JsVariableDeclarationClause` that has an optional semicolon (same as `VariableStatement`). 

Can we add the semicolon to the `VariableDeclarations` node? No, I tried that. The problem is that `VariableDeclarations` are also valid inside the initialiser of a `for (let a, b;;) {}` loop and a semicolon isn't allowed in that case (or the parser eats the for separator semicolons. 

This PR further re-introduces the `TsDeclareStatement` for all `declare JsAnyDeclarationClause` declarations and adds `export type`, and `export enum` support.