chore(coverage): Add `--detailed` option to `cargo coverage` (#2112)
The coverage reporter prints the progress with all ran tests to `stderr`. This means, that it isn't possible to pipe the results to a text file.

This PR adds a `--detailed` option that allows printing the failing tests with their diagnostics after running all tests to `stdout` (except if `--json` is passed). This allows piping the test results to a text file for analysis.

This PR further changes:

* `--json`: Forces that all other reporters print to `stderr` instead of `stdout`. It's now possible to use all options in combinations
* `--filter=<file>` Introduce the filter option that replaces the previous free argument
* `--detailed=rast`: Replaces `--show-rast`. It prints the rast of all failing tests in the summary

I updated the documentation to reflect the changes.

I had to fix a bug in the diagnostics crate related to incorrectly indexing into the middle of characters. The issue is that the unicode line breaks are longer than 1 byte. Thus, they break
the `+ 1` arithmetic used throughout the diagnostics package to get the next character right before/after a line break.

I don't think we need to handle these line breaks as "proper" line breaks because many editors
fall back to render the unicode line breaks as a special character.