fix(rslint_parser): Type-arguments in `extends` heritage clause
The parser incorrectly parsed `B<A>` in `class T extends B<A>` as the start of a call expression with type arguments (assumed it would be `B<A>()`) instead of an identifier with type arguments.

This PR adds a check when speculatively parsing the type arguments of a call expression to only eat them if they're followed by a token that is valid coming after type arguments (either a template literal or `(`.

We may need to extend the list of valid tokens in the future to get better error recovery.