fix(rslint_errors): Fix line numbers for files using `\r\n`
The error packages determines the line numbers by searching the positions of the `\n` and `\r` characters and taking them as the start of a new line.
The problem with this approach is that it creates two lines for a `\r\n`: It first assumes a line break after `\r` and then after `\n`.

This PR fixes the line counting for `\r\n` by testing if a `\r` is followed by a `\n` and if so, don't treat it as a line break (only the following `\n`).