fix(rslint_parser): Stop statement and module recovery before closing `}`
Example:
```
{
"name": "troublesome-lib",
"typings": "lib/index.d.ts",
"version": "0.0.1"
}
```

The parser tries to parse this as a block statement but then fails because of the `"name":`. The existing error recovery incorrectly eats away all tokens including `}` which then results in a parse error that the closing `}` is missing.

Before:

```
error[SyntaxError]: Expected a semicolon or an implicit semicolon after a statement, but found none
  ┌─ statements_closing_curly.js:3:7
  │
3 │ "name": "troublesome-lib",
  │ ------^
  │ │     │
  │ │     An explicit or implicit semicolon is expected here...
  │ ...Which is required to end this statement

--
error[SyntaxError]: expected `'}'` but instead the file ends
  ┌─ statements_closing_curly.js:7:1
  │
7 │
  │ ^ the file ends here
```

After:

```
error[SyntaxError]: Expected a semicolon or an implicit semicolon after a statement, but found none
  ┌─ statements_closing_curly.js:2:7
  │
2 │ "name": "troublesome-lib",
  │ ------^
  │ │     │
  │ │     An explicit or implicit semicolon is expected here...
  │ ...Which is required to end this statement

```