fix(rslint_parser): Move `arrow` parsing into `parse_assignment_expression_or_higher` (#2136)
Arrow functions are only allowed in [`AssignmentExpression` positions](https://tc39.es/ecma262/multipage/ecmascript-language-expressions.html#sec-assignment-operators) but the current implementation also parses out arrow functions in `PrimaryExpression` positions. 

This PR moves the parsing of arrow functions into the `parse_primary_expression` function. What makes parsing arrow function expressions challenging is that the head of an arrow function (everything before the `=>`) can as well be a parenthesized expression in many cases. For example:

```
(a, b, c, { d }) // parenthesized
(a, b, c, { d}) => {} // arrow function, because of the `{}`
```

This ambiguity means, the parser must speculatively parse the whole syntax between `(` and `)` to determine if it is a valid arrow function and, if not, re-parse it as a parenthesized expression. This works reasonably well in is also what TS is doing but causes issues if, for example, the parsing of a parameter's binding encounters any invalid syntax and, thus, error recovery kicks in:

```js
(a, 5 + 3) 
```

The parser will try to perform error recovery if this parenthesized expression is parsed as an arrow function because `5 + 3` isn't a valid binding. Our recovery is even good enough that the parser correctly recovers at the `)` which then gives the impression that the parser was able to correctly parse out the arrow function. 

This PR re-introduces the `speculative_parsing` state (previously called `no_recovery`) to disable error recovery when speculatively parsing any expression. Completely disabling error recovery means that our arrow function parameter recovery may not be as good as it could be because error recovery doesn't kick in. This is something that we can further improve in the future if it proves to be a problem. 

An alternative to this implementation would have been to introduce a new `ExpressionContext` flag that indicates if the current expression should be parsed as an arrow function or not. The downside I see there is that 

1. Requires an additional flag whereas moving it into the `parse_assignment_expression_or_higher` gives us the information for free
2. Parsing sub-expressions inherits the `ExpressionContext`. It would, therefore, have been required to reset the `allow_paren` flag for all calls parsing out sub-expressions that don't call into `parse_assignment_expression_or_higher`. Again, this is something we get for free when moving it into the appropriate function. 

Another benefit of the current implementation is that it moves the parsing of `arrow` functions into a single place. The parsing logic used to be spread all over the expression parsing: `async` functions, `type arguments`, parenthesized or not. Unifying the logic fixed issues with `async` functions with type arguments not being correctly parsed and the parser ignoring line breaks before `=>`