fix(rslint_parser): TS `asserts` predicate
TypeScript supports two kind of type refinement functions:

**[Type Predicates]**

```ts
function isString(a): a is string {
  return typeof a === "string"
}

function test(a) {
  if isString(a) {
    console.log("a is a string"); // The type of `a` is string
  }
}
```

The function returns a boolean whatever the specified argument fulfills the condition.

**[Type Assertions](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#assertion-functions)

```ts
function assertString(a): asserts a is string {
  if typeof a !== "string" {
    throw new Error("A is supposed to be a string");
  }
}

function test(a) {
  assertString(a);
  console.log(a); // The type of `a` is string
}
```

Asserts are in that sense special that they also allow omitting the `is ...` part.

```ts
function asserts(condition: any): asserts condition {
  if (!condition) {
    throw new Error("Whoops");
  }
}

function test(a) {
  asserts(a > 10);
  // a is guaranteed to be bigger than 10 and has the type `number`
}
```

This second part is something that our current AST didn't model correctly

This PR introduces a new `TsAssertsReturnType` for `asserts...` refinements and renames `TsTypePredicate` to `TsPredicateReturnType` because these are types (they appear in type positions).